Massively simplified (and sped up) the code. No longer allocating a large chunk of memory to write in and then write to the strip.
All writing to the LED strip happens directly from the function that calculates the wcolor wave. This means that there's no time lost in buffering all the color information to a chunk of memory, and then writing that chunk of memory to the strip. All sending of information is done using plain SPI.transfer() methods instead of "spooky" writing to SPDR and waiting for other registers to clear.