Complete rewrite of BouncyBallPhysics.
Rewrote the core of the BouncyBallPhysics sketch to a Ball class rather than just a struct to hold the data. This allows all the movement to be handled by methods of the ball, gravity working on it and the ball moving as a result of it. This allows for many falling and bouncing objects to be processed on the strip.

Additionally, the code that handles the bounce has been improved to calculate the exact moment of the bounce, placing the ball in the correct spot after the bounce (with the correct speed and direction).

Some additional cleanup and rewriting needs to be done to complete the class. There also needs to be a slightly better color management, as multiple balls are difficult to separate from each other at the moment. A color that changes intensity seems likely.