Fixed timeForDisplacement where acceleration == 0
This _actually_ resolves #1 (the previous commit only fixed the negative gravity case, not the negative gravity

Also swapped a number of duration,acceleration arguments, the duration coming first and the acceleration second. Renamed the 'accelerate' method on the ball to 'travel' that takes an optional acceleration argument, the default of which is zero. Zero gravity simulations now work properly :)