Simplified RunningColorDot
Reduced the complexity of the program (and possible small visual errors) by reducing the number of 'timed' events to one (the change of color). Progression of the trail is now expressed in number of color changes.

This ensures that the changing of color and the trail movement are factors of eachother, which prevents problems where the trail moves in a jerky fashion because of aliasing between the movement and the painting of it.