Added RedRunner project.
This paints a long monochrome (red) trail of about 16 pixels that moves along the LED-strip similar to the RunningColorDot. The brightness of the trail is corrected for non-linear luminance perception (based on CIELAB models). This is the first step in the project, more additions are to follow.

Also improved an explanation on the RunningColorDot code.