Added barriers to the RedRunner demo.
Every 150ms, there is a one in four chance of a barrier appearing. These barriers interact with the running trail: whenever the trail runs into a barriers, the direction of the trail changes. The barrier also loses 'structural integrity' (brightness) as a result of the collision. After 4 collisions the barrier is destroyed (no longer lit up) and no longer interacts with the trail.